Changed: dedupe agent-runtime test fixtures and drop redundant tests
Audit of task.rs's test module found two tests asserting nothing their
neighbours did not already cover (public-wrapper variants of the
no-callable-targets and max-depth scenarios) and four copies of the same
fixtures spread across task.rs, handle.rs, build.rs, and the hook
examples. This removes the redundancy without losing executed coverage:

- Deleted `agent_build_context_omits_task_tool_when_no_targets_are_callable`
  (byte-identical scenario to `build_agent_skips_task_tool_when_no_targets_
  are_callable`; the only delta, the 1-line `build()` delegate, stays
  covered by the hook end-to-end test) and
  `agent_build_context_omits_task_tool_when_max_depth_is_zero` (same
  production branch as `build_agent_omits_task_tool_at_max_depth`).
- Merged the pattern-scoped and absent-permission Task-attach tests into
  one two-runtime test; both assertion sets remain.
- Promoted `two_tools_then_text` into `mock` as a public generalisation of
  `tool_then_text` (shared `tool_call_response` helper, reuses
  `extract_tool_return_text`), replacing the test-local copy and the twin
  in `examples/hooks/shared.rs`; the tool-block example now imports it.
- Moved agent/allow_tools/pattern_task/catalog/credentials/workspace_root
  fixtures into `agent_runtime::test_stubs` (pub(crate), cfg(test));
  handle.rs and build.rs test modules consume them and build.rs's
  `agent_with_sampling` becomes a struct update on the shared fixture.
- Replaced cfg-gated `TaskBuildContext` struct literals with the existing
  `new_for_test` constructor, decoupling tests from future field
  additions.

Net -285 lines. Suite stays green (110 unit + 20 doc tests); the hook
end-to-end test still solely covers `with_model_override`,
`HookedAgent::run`, and the `HookedToolExecutor` wiring.
