+++ src/reloaded-code-serdesai/src/mock.rs
@@ -155,7 +155,6 @@ pub fn tool_then_text(
 ) -> Streamed<FunctionModel> {
     let tool_name = tool_name.into();
     let fallback_text = fallback_text.into();
-    let tool_name_clone = tool_name.clone();
 
     let model = FunctionModel::new(move |messages, _settings| {
         // Check whether the conversation already contains a tool return from a
@@ -185,15 +184,85 @@ pub fn tool_then_text(
             ModelResponse::text(text)
         } else {
             // First call: emit a tool call so the agent executes the real tool.
-            ModelResponse::with_parts(vec![
-                ModelResponsePart::text(format!("Calling {tool_name}...")),
-                ModelResponsePart::tool_call(tool_name_clone.clone(), args.clone()),
-            ])
-            .with_finish_reason(FinishReason::ToolCall)
+            tool_call_response(&tool_name, &args)
         }
     });
 
     Streamed::new(model)
+}
+
+/// Scripts two tool calls followed by a text answer, generalising
+/// [`tool_then_text`] to two-step flows: the first turn calls `first`, the
+/// next turn calls `second`, and once both tool returns are in the
+/// conversation the model answers with `final_text` followed by the real
+/// tool returns, so callers can observe what the model received.
+///
+/// # Arguments
+///
+/// - `first` - name and JSON arguments of the tool called on the first turn.
+/// - `second` - name and JSON arguments of the tool called on the second turn.
+/// - `final_text` - text prefix for the closing response; the real tool
+///   returns are appended after it.
+///
+/// # Example
+///
+/// ```rust,no_run
+/// use reloaded_code_serdesai::mock::two_tools_then_text;
+/// use serde_json::json;
+///
+/// let model = two_tools_then_text(
+///     ("read", json!({"file_path": "service.env"})),
+///     ("write", json!({"file_path": "draft.md", "content": "notes"})),
+///     "Run finished.",
+/// );
+/// ```
+pub fn two_tools_then_text(
+    first: (impl Into<String>, serde_json::Value),
+    second: (impl Into<String>, serde_json::Value),
+    final_text: impl Into<String>,
+) -> Streamed<FunctionModel> {
+    let first_name = first.0.into();
+    let first_args = first.1;
+    let second_name = second.0.into();
+    let second_args = second.1;
+    let final_text = final_text.into();
+
+    let model = FunctionModel::new(move |messages, _settings| {
+        // Each finished tool call adds one tool return to the history, so
+        // their count tells which scripted turn is next.
+        let answered_calls = messages.iter().flat_map(|m| m.tool_returns()).count();
+
+        match answered_calls {
+            0 => tool_call_response(&first_name, &first_args),
+            1 => tool_call_response(&second_name, &second_args),
+            _ => {
+                let tool_results: String = messages
+                    .iter()
+                    .flat_map(|m| m.tool_returns())
+                    .map(extract_tool_return_text)
+                    .collect::<Vec<_>>()
+                    .join("\n");
+                let text = if tool_results.is_empty() {
+                    final_text.clone()
+                } else {
+                    format!("{final_text}\n\n{tool_results}")
+                };
+                ModelResponse::text(text)
+            }
+        }
+    });
+
+    Streamed::new(model)
+}
+
+/// Emits a tool-call response: a short text part, then the call that
+/// triggers the real tool.
+fn tool_call_response(tool_name: &str, args: &serde_json::Value) -> ModelResponse {
+    ModelResponse::with_parts(vec![
+        ModelResponsePart::text(format!("Calling {tool_name}...")),
+        ModelResponsePart::tool_call(tool_name, args.clone()),
+    ])
+    .with_finish_reason(FinishReason::ToolCall)
 }
 
 /// Extract human-readable text from a [`ToolReturnPart`].
